move and update test_stack_sets
