--- tests/aws/services/cloudformation/v2/ported_from_v1/resources/test_stack_sets.snapshot.json
@@ -1,32 +1,0 @@
-{
-  "tests/aws/services/cloudformation/v2/ported_from_v1/resources/test_stack_sets.py::test_create_stack_set_with_stack_instances": {
-    "recorded-date": "24-05-2023, 15:32:47",
-    "recorded-content": {
-      "create_stack_set": {
-        "StackSetId": "<stack-set-id:1>",
-        "ResponseMetadata": {
-          "HTTPHeaders": {},
-          "HTTPStatusCode": 200
-        }
-      },
-      "create_stack_instances": {
-        "OperationId": "<uuid:1>",
-        "ResponseMetadata": {
-          "HTTPHeaders": {},
-          "HTTPStatusCode": 200
-        }
-      }
-    }
-  },
-  "tests/aws/services/cloudformation/v2/ported_from_v1/resources/test_stack_sets.py::test_fetch_non_existent_stack_set_instances": {
-    "recorded-date": "25-07-2025, 14:15:09",
-    "recorded-content": {
-      "non-existent-stack-set-name": "An error occurred (StackSetNotFoundException) when calling the CreateStackInstances operation: StackSet non-existent-stack-set-id not found",
-      "non-existent-stack-set-id": "An error occurred (StackSetNotFoundException) when calling the CreateStackInstances operation: StackSet <stack-id> not found"
-    }
-  },
-  "tests/aws/services/cloudformation/v2/ported_from_v1/resources/test_stack_sets.py::test_delete_nonexistent_stack_set": {
-    "recorded-date": "25-07-2025, 14:57:22",
-    "recorded-content": {}
-  }
-}
